backend/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
coverage==7.6.12
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
